save history of router
