+++ src/state/actions/router.js
@@ -1,6 +1,6 @@
 import { createAction } from "redux-actions"
 
-import { CHANGE_VIEW } from "../types"
+import { CHANGE_VIEW, BACK_TO_PREVIOUS_VIEW } from "../types"
 
-// eslint-disable-next-line
 export const changeView = createAction(CHANGE_VIEW)
+export const backToPreviousView = createAction(BACK_TO_PREVIOUS_VIEW)
